Address connector migration review comments
