+++ internal/core/workflow_migrate_connection_version_types.go
@@ -1,0 +1,60 @@
+package core
+
+import (
+	"github.com/rmorlok/authproxy/internal/database"
+	cschema "github.com/rmorlok/authproxy/internal/schema/resources/connectors"
+)
+
+// migrationHookPatch is the normalized object returned by connector-authored
+// migration JavaScript. It records the configuration, label, annotation, and
+// notification changes that should be applied after a hook runs successfully.
+type migrationHookPatch struct {
+	Config        migrationAnyPatch          `json:"config"`
+	Labels        migrationStringPatch       `json:"labels"`
+	Annotations   migrationStringPatch       `json:"annotations"`
+	Notifications []migrationNotificationDef `json:"notifications"`
+}
+
+// migrationAnyPatch describes set/unset operations for JSON-like maps such as
+// connection configuration, where values may be strings, numbers, objects, or
+// arrays.
+type migrationAnyPatch struct {
+	Set   map[string]any `json:"set"`
+	Unset []string       `json:"unset"`
+}
+
+// migrationStringPatch describes set/unset operations for string-valued maps
+// such as connection labels and annotations.
+type migrationStringPatch struct {
+	Set   map[string]string `json:"set"`
+	Unset []string          `json:"unset"`
+}
+
+// migrationNotificationDef is the connector-authored notification payload that
+// a hook can queue for users after the connection version changes.
+type migrationNotificationDef struct {
+	Key       string         `json:"key"`
+	Level     string         `json:"level"`
+	Title     string         `json:"title"`
+	Message   string         `json:"message"`
+	ActionURL string         `json:"action_url"`
+	Metadata  map[string]any `json:"metadata"`
+}
+
+// connectionMigrationCandidate is the in-memory target state assembled by hook
+// execution and automatic migration analysis before the workflow writes the
+// connection update and notification changes to persistent storage.
+type connectionMigrationCandidate struct {
+	Connection       *connection
+	Target           *ConnectorVersion
+	Config           map[string]any
+	UserLabels       map[string]string
+	Annotations      map[string]string
+	SetupStep        *cschema.SetupStep
+	SetupError       *string
+	HealthState      database.ConnectionHealthState
+	RefreshAuth      bool
+	ProbeIdsToRun    []string
+	Notifications    []database.NotificationUpsert
+	NotificationKeys []string
+}
